feat: solve cross domain issues with other small fixes

+++ backend/app/routes/user.py
@@ -13,11 +13,13 @@ def load_user(user_id):
     return User.query.get(int(user_id))
 
 
-@bp.route('/logout', methods=['POST'])
-@login_required
+@bp.route('/logout', methods=['DELETE'])
 def logout():
-    logout_user()
-    return jsonify({"message": "退出登录成功"}), 200
+    if current_user.is_authenticated:
+        logout_user()
+        return jsonify({'success': True, 'message': '退出登录成功'})
+    else:
+        return jsonify({'success': False, 'message': '用户未登录'})
 
 
 @bp.route('/login', methods=['GET', 'POST'])
@@ -26,7 +28,7 @@ def login():
         return redirect(url_for('main.index'))
 
     if current_user.is_authenticated:
-        return jsonify({'message': '用户已登录'}), 401
+        return jsonify({'success': False, 'message': '用户已登录'})
 
     form = LoginForm()
     if form.validate():
@@ -38,11 +40,11 @@ def login():
         if user and user.check_password(password):
             login_user(user)
             user.update_last_login()
-            return jsonify({'message': '登录成功'}), 200
+            return jsonify({'success': True, 'message': '登录成功', 'user': user.to_dict()})
         else:
-            return jsonify({'message': '用户名或密码不正确'}), 401
+            return jsonify({'success': False, 'message': '用户名或密码不正确'})
     else:
-        return jsonify({'message': form.errors}), 401
+        return jsonify({'success': False, 'message': form.errors})
 
 
 @bp.route('/signup', methods=['GET', 'POST'])
@@ -59,34 +61,40 @@ def signup():
         new_user.set_password(form.password.data)
         db.session.add(new_user)
         db.session.commit()
-        return jsonify({'message': '注册成功'}), 201
+        return jsonify({'success': True, 'message': '注册成功'})
     else:
-        return jsonify({'message': form.errors}), 400
+        return jsonify({'success': False, 'message': form.errors})
 
 
 @bp.route('/unregister', methods=['DELETE'])
 def unregister():
-    if current_user.is_authenticated:
-        try:
-            # 删除与当前用户相关的 price_alert 记录
-            PriceAlert.query.filter_by(user_id=current_user.id).delete()
-            # 删除当前用户
-            db.session.delete(current_user)
-            db.session.commit()
-            return jsonify({'message': '注销成功'}), 200
-        except IntegrityError as e:
-            db.session.rollback()
-            return jsonify({'message': '注销失败，数据库操作失败', 'error': str(e)}), 500
-    else:
-        return jsonify({'message': '用户未登录'}), 401
+    if not current_user.is_authenticated:
+        return jsonify({'success': False, 'message': '用户未登录'})
+    try:
+        # 删除与当前用户相关的 price_alert 记录
+        PriceAlert.query.filter_by(user_id=current_user.id).delete()
+        # 删除当前用户
+        db.session.delete(current_user)
+        db.session.commit()
+        return jsonify({'success': True, 'message': '注销成功'})
+    except IntegrityError as e:
+        db.session.rollback()
+        return jsonify({'success': False, 'message': '注销失败，数据库操作失败', 'error': str(e)})
+
+
+@bp.route('/verify', methods=['GET'])
+def verify():
+    return jsonify({'success': current_user.is_authenticated})
 
 
 @bp.route('/alert', methods=['POST'])
-@login_required
 def set_alert():
+    if not current_user.is_authenticated:
+        return jsonify({'success': False, 'message': '用户未登录'})
+
     product = request.json
     if PriceAlert.query.filter_by(user_id=current_user.id, product_id=product['id']).first():
-        return jsonify({'message': '价格提醒已设置'}), 400
+        return jsonify({'success': False, 'message': '价格提醒已设置'})
 
     new_alert = PriceAlert(
         user_id=current_user.id,
@@ -95,24 +103,23 @@ def set_alert():
     )
     db.session.add(new_alert)
     db.session.commit()
-    return jsonify({'message': '价格提醒设置成功'}), 201
-
-
-@bp.route('/alert', methods=['GET'])
-@login_required
-def get_alerts():
-    alerts = PriceAlert.query.filter_by(user_id=current_user.id).all()
-    return jsonify([alert.to_dict() for alert in alerts]), 200
+    return jsonify({'message': '价格提醒设置成功'})
 
 
 @bp.route('/alert', methods=['DELETE'])
-@login_required
 def delete_alert():
+    if not current_user.is_authenticated:
+        return jsonify({'success': False, 'message': '用户未登录'})
+
     product = request.json
     alert = PriceAlert.query.filter_by(user_id=current_user.id, product_id=product['id']).first()
     if not alert:
-        return jsonify({'message': '价格提醒不存在'}), 404
+        return jsonify({'success': False, 'message': '价格提醒不存在'})
 
-    db.session.delete(alert)
-    db.session.commit()
-    return jsonify({'message': '价格提醒删除成功'}), 200
+    try:
+        db.session.delete(alert)
+        db.session.commit()
+        return jsonify({'success': True, 'message': '价格提醒删除成功'})
+    except Exception as e:
+        db.session.rollback()
+        return jsonify({'success': False, 'message': '价格提醒删除失败', 'error': str(e)})
